pyproject.toml
[project]
name = "pgfirstaid-testing"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "psycopg2-binary>=2.9.12",
    "pytest>=8.0",
]

[tool.pytest.ini_options]
testpaths = ["testing"]
